fixed state problem with the gameState : added a useRef to keep gameState updated inside the customHook and a useState to use the gameState outside the hook
